extract resolveProxyHost to a module. (#184)
